+++ tests/test_rosenbrock_chain.py
@@ -18,7 +18,7 @@ def test_run_d40_rosen():
     allowable_evals = (52.2 + 1.5*3) * 1e3 # 2 sigma
     iteration_number = int(allowable_evals / lamb) + 1
 
-    cr = CRFMNES(dim, rosenbrock_chain, mean, sigma, lamb)
+    cr = CRFMNES(dim, rosenbrock_chain, mean, sigma, lamb, dtype=np.float128)
     x_best, f_best = cr.optimize(iteration_number)
     print("f_best:{}".format(f_best))
     assert f_best < 1e-12
@@ -33,7 +33,7 @@ def test_run_d80_rosen():
     allowable_evals = (192 + 2.0*3) * 1e3 # 2 sigma
     iteration_number = int(allowable_evals / lamb) + 1
 
-    cr = CRFMNES(dim, rosenbrock_chain, mean, sigma, lamb)
+    cr = CRFMNES(dim, rosenbrock_chain, mean, sigma, lamb, dtype=np.float128)
     x_best, f_best = cr.optimize(iteration_number)
     print("f_best:{}".format(f_best))
     assert f_best < 1e-12
